Fixed the issue with the tooltip and added code to InventoryController class for consistency.
