+++ Assets/Code/GameManager.cs
@@ -33,11 +33,6 @@ namespace Code.Inventory
         public static event Action<GameState> OnGameStateChanged;
         // Ending Screen
         public GameObject endingScreen;
-        // Inventory items
-        public List<Item> _itemList = new List<Item>();
-        public Transform _inventoryPanel;
-        public GameObject _itemInfoPrefab;
-        private GameObject _currentItemInfo = null;
 
         [NonSerialized] public bool IsGameOver;
         [NonSerialized] public bool IsEndOfChapter;
@@ -53,23 +48,6 @@ namespace Code.Inventory
             NewGame,
             LoadGame
         }
-
-        // Singleton region to destroy items as well as item description
-        #region singleton
-
-        public static GameManager instance;
-
-        private void Awake()
-        {
-            if (Gm == null)
-                Gm = this;
-            else
-                Destroy(gameObject);
-
-            if (instance == null)
-                instance = this;
-        }
-        #endregion
 
         private void Start()
         {
@@ -133,10 +111,6 @@ namespace Code.Inventory
                 LoadNextStoryPart();
             if (IsGameOver)
                 LoadGameOverScreen();
-            // Just to test is the rest of the system with the inventory works fine so far
-            // TODO: Make sure to implement system for actually adding items through nodes
-            //if (Input.GetKeyDown(KeyCode.X))
-                //Inventory._instance.AddItem(_itemList[Random.Range(0, _itemList.Count)]);
         }
 
         /// <summary>
@@ -202,83 +176,5 @@ namespace Code.Inventory
             }
             return 0;
         }
-
-        // Made this region to better distinguish between Item code and actual GameManager
-        #region ItemRegion
-
-        
-        /// <summary>
-        /// Adds the Item,
-        /// TODO Put into InventoryController
-        /// </summary>
-        /// <param name="item"></param>
-        public void AddItem(string item)
-        {
-            foreach (var i in _itemList)
-            {
-                if (i._name.Equals(item))
-                {
-                    Inventory._instance.AddItem(i);
-                }
-            }
-        }
-        
-        /// <summary>
-        /// This function is called when we use an item that can grant stats, such as health.
-        /// </summary>
-        /// <param name="itemType"></param>
-        /// <param name="amount"></param>
-        public void OnStatItemUse(StatItemType itemType, int amount)
-        {
-            Debug.Log("Consuming: "+ itemType + " Added amount: "+amount);
-        }
-        
-        /// <summary>
-        /// Displays item information when the mouse is hovered over an item.
-        /// </summary>
-        /// <param name="itemName"></param>
-        /// <param name="itemDescription"></param>
-        /// <param name="buttonPos"></param>
-        public void DisplayItemInfo(string itemName, string itemDescription, Vector2 buttonPos)
-        {
-            if (_currentItemInfo != null)
-            {
-                Destroy(_currentItemInfo.gameObject);
-            }
-        
-            // Create variables so that the tooltip window is moved to a different direction and not simply closed
-            buttonPos.x -= 80;  
-            buttonPos.y += 40;
-            
-            _currentItemInfo = Instantiate(_itemInfoPrefab, buttonPos, Quaternion.identity, _inventoryPanel);
-            _currentItemInfo.GetComponent<ItemInfo>().SetUp(itemName, itemDescription);
-        }
-        
-        /// <summary>
-        /// Destroy Item info so it's not staying behind when the mouse is leaving.
-        /// </summary>
-        public void DestroyItemInfo()
-        {
-            if (_currentItemInfo != null)
-            {
-                float delayInSeconds = 0.5f; // Set the delay in seconds
-        
-                // Delay the destruction of the _currentItemInfo game object
-                StartCoroutine(DelayedDestroy(_currentItemInfo.gameObject, delayInSeconds));
-        
-                _currentItemInfo = null; // Clear the _currentItemInfo variable
-            }
-        }
-
-        IEnumerator DelayedDestroy(GameObject gameObjectToDestroy, float delayInSeconds)
-        {
-            // Wait for the specified delay
-            yield return new WaitForSeconds(delayInSeconds);
-            
-            // Destroy the game object
-            Destroy(gameObjectToDestroy);
-        }
-
-        #endregion
     }
 }
