+++ Assets/Code/Inventory/InventoryController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9d9548a93c5a9244ebfb0afb5fbe6f2d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
