+++ Assets/Code/Logger/Log Files/GameManager.log
@@ -4,3 +4,6 @@ Line Number: 79
 LogTime = 4/3/2023 8:19:38 PM | LogType: GameManager Log 
 LogMessage: Game Over!  
 Line Number: 176
+LogTime = 4/3/2023 10:12:15 PM | LogType: GameManager Log 
+LogMessage: Next chapter: Story1Chapter2 
+Line Number: 155
