+++ Assets/Code/Logger/Log Files/StoryEditor.log.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 806bba4146e0d814aa17209f2a6a5238
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
