+++ Assets/Code/Logger/Log Files/StoryHolder.log
@@ -1,3 +1,84 @@
-LogTime = 4/3/2023 8:19:37 PM | LogType: Story Holder log 
-LogMessage: Returning next Choice node dd042f74-2a64-4461-a8ea-9af4044edf11 
-Line Number: 72
+LogTime = 4/5/2023 7:56:37 AM | LogType: Story Holder log 
+LogMessage: Returning next Choice node 3c10b112-d543-43a3-ae9d-ccdfca1e4fb2 
+Line Number: 73
+LogTime = 4/5/2023 8:03:24 AM | LogType: Story Holder log 
+LogMessage: Returning next Choice node 3c10b112-d543-43a3-ae9d-ccdfca1e4fb2 
+Line Number: 73
+LogTime = 4/5/2023 8:05:40 AM | LogType: Story Holder log 
+LogMessage: Returning next Choice node 3c10b112-d543-43a3-ae9d-ccdfca1e4fb2 
+Line Number: 73
+LogTime = 4/5/2023 8:06:32 AM | LogType: Story Holder log 
+LogMessage: Returning next Choice node 3c10b112-d543-43a3-ae9d-ccdfca1e4fb2 
+Line Number: 73
+LogTime = 4/5/2023 8:06:55 AM | LogType: Story Holder log 
+LogMessage: Returning next Choice node 3c10b112-d543-43a3-ae9d-ccdfca1e4fb2 
+Line Number: 73
+LogTime = 4/5/2023 8:07:05 AM | LogType: Story Holder log 
+LogMessage: Returning next Choice node 3c10b112-d543-43a3-ae9d-ccdfca1e4fb2 
+Line Number: 73
+LogTime = 4/5/2023 8:16:36 AM | LogType: Story Holder log 
+LogMessage: Returning next Choice node 3c10b112-d543-43a3-ae9d-ccdfca1e4fb2 
+Line Number: 73
+LogTime = 4/5/2023 8:16:40 AM | LogType: Story Holder log 
+LogMessage: Returning next Choice node 6777389a-b747-4241-8a61-88941f6d62ac 
+Line Number: 73
+LogTime = 4/5/2023 8:16:41 AM | LogType: Story Holder log 
+LogMessage: Returning next Story node a50d2f53-367c-4c8c-b69a-71d802763b8f 
+Line Number: 87
+LogTime = 4/5/2023 8:16:46 AM | LogType: Story Holder log 
+LogMessage: Returning next Choice node 3c10b112-d543-43a3-ae9d-ccdfca1e4fb2 
+Line Number: 73
+LogTime = 4/5/2023 10:26:42 AM | LogType: Story Holder log 
+LogMessage: Returning next Choice node 3c10b112-d543-43a3-ae9d-ccdfca1e4fb2 
+Line Number: 73
+LogTime = 4/5/2023 10:27:39 AM | LogType: Story Holder log 
+LogMessage: Returning next Choice node 3c10b112-d543-43a3-ae9d-ccdfca1e4fb2 
+Line Number: 73
+LogTime = 4/10/2023 12:37:47 PM | LogType: Story Holder log 
+LogMessage: Returning next Choice node 3c10b112-d543-43a3-ae9d-ccdfca1e4fb2 
+Line Number: 73
+LogTime = 4/10/2023 12:37:55 PM | LogType: Story Holder log 
+LogMessage: Returning next Choice node 3c10b112-d543-43a3-ae9d-ccdfca1e4fb2 
+Line Number: 73
+LogTime = 4/10/2023 1:17:19 PM | LogType: Story Holder log 
+LogMessage: Returning next Choice node 3c10b112-d543-43a3-ae9d-ccdfca1e4fb2 
+Line Number: 73
+LogTime = 4/10/2023 1:18:16 PM | LogType: Story Holder log 
+LogMessage: Returning next Choice node 3c10b112-d543-43a3-ae9d-ccdfca1e4fb2 
+Line Number: 73
+LogTime = 4/10/2023 1:45:45 PM | LogType: Story Holder log 
+LogMessage: Returning next Choice node 3c10b112-d543-43a3-ae9d-ccdfca1e4fb2 
+Line Number: 73
+LogTime = 4/10/2023 1:46:06 PM | LogType: Story Holder log 
+LogMessage: Returning next Choice node 3c10b112-d543-43a3-ae9d-ccdfca1e4fb2 
+Line Number: 73
+LogTime = 4/10/2023 4:21:30 PM | LogType: Story Holder log 
+LogMessage: Returning next Choice node 3c10b112-d543-43a3-ae9d-ccdfca1e4fb2 
+Line Number: 73
+LogTime = 4/10/2023 4:21:33 PM | LogType: Story Holder log 
+LogMessage: Returning next Choice node 6777389a-b747-4241-8a61-88941f6d62ac 
+Line Number: 73
+LogTime = 4/10/2023 4:22:11 PM | LogType: Story Holder log 
+LogMessage: Returning next Choice node 3c10b112-d543-43a3-ae9d-ccdfca1e4fb2 
+Line Number: 73
+LogTime = 4/10/2023 4:23:29 PM | LogType: Story Holder log 
+LogMessage: Returning next Choice node 3c10b112-d543-43a3-ae9d-ccdfca1e4fb2 
+Line Number: 73
+LogTime = 4/10/2023 4:23:44 PM | LogType: Story Holder log 
+LogMessage: Returning next Choice node 3c10b112-d543-43a3-ae9d-ccdfca1e4fb2 
+Line Number: 73
+LogTime = 4/10/2023 4:25:46 PM | LogType: Story Holder log 
+LogMessage: Returning next Choice node 3c10b112-d543-43a3-ae9d-ccdfca1e4fb2 
+Line Number: 73
+LogTime = 4/10/2023 4:28:44 PM | LogType: Story Holder log 
+LogMessage: Returning next Choice node 3c10b112-d543-43a3-ae9d-ccdfca1e4fb2 
+Line Number: 73
+LogTime = 4/10/2023 4:29:20 PM | LogType: Story Holder log 
+LogMessage: Returning next Choice node 3c10b112-d543-43a3-ae9d-ccdfca1e4fb2 
+Line Number: 73
+LogTime = 4/10/2023 4:30:12 PM | LogType: Story Holder log 
+LogMessage: Returning next Choice node 3c10b112-d543-43a3-ae9d-ccdfca1e4fb2 
+Line Number: 73
+LogTime = 4/10/2023 4:30:34 PM | LogType: Story Holder log 
+LogMessage: Returning next Choice node 3c10b112-d543-43a3-ae9d-ccdfca1e4fb2 
+Line Number: 73
