+++ Assets/Code/Logger/Log Files/StoryUI.log
@@ -1,3 +1,3 @@
-LogTime = 4/3/2023 8:19:38 PM | LogType: UI log 
+LogTime = 4/10/2023 4:21:19 PM | LogType: UI log 
 LogMessage: Game Over reached. 
-Line Number: 195
+Line Number: 208
