+++ Assets/Prefabs/GameUI/Items/ItemPrefab/ItemInfoPrefab.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 65c9db9d143c23a408eb0dff1c83d30b
+guid: 86df1762ca180e246a48e94cdc92a659
 PrefabImporter:
   externalObjects: {}
   userData:
