+++ Assets/Scenes/Part1Scene.unity
@@ -159,9 +159,9 @@ RectTransform:
   m_RootOrder: 2
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 1, y: 0}
-  m_AnchorMax: {x: 1, y: 1}
+  m_AnchorMax: {x: 1, y: 0}
   m_AnchoredPosition: {x: 0, y: 0}
-  m_SizeDelta: {x: 20, y: -17}
+  m_SizeDelta: {x: 20, y: 0}
   m_Pivot: {x: 1, y: 1}
 --- !u!114 &72947665
 MonoBehaviour:
@@ -286,7 +286,7 @@ RectTransform:
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 1, y: 1}
   m_AnchoredPosition: {x: 0, y: 0.00037838938}
-  m_SizeDelta: {x: 0, y: 20}
+  m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0, y: 1}
 --- !u!114 &108057771
 MonoBehaviour:
@@ -362,9 +362,9 @@ RectTransform:
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
-  m_AnchorMax: {x: 1, y: 1}
+  m_AnchorMax: {x: 0, y: 0}
   m_AnchoredPosition: {x: 0, y: 0}
-  m_SizeDelta: {x: -17, y: -17}
+  m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0, y: 1}
 --- !u!114 &118720622
 MonoBehaviour:
@@ -417,6 +417,72 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 118720620}
   m_CullTransparentMesh: 1
+--- !u!1 &186694792
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 186694795}
+  - component: {fileID: 186694794}
+  - component: {fileID: 186694793}
+  m_Layer: 0
+  m_Name: InventoryController
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &186694793
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 186694792}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 9d9548a93c5a9244ebfb0afb5fbe6f2d, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _itemList:
+  - {fileID: 11400000, guid: 08c594da8b028ca44b156b2b718a738f, type: 2}
+  - {fileID: 11400000, guid: b793a0ba0d49da840a4eef3ba31165d6, type: 2}
+  - {fileID: 11400000, guid: 9feb9742a522f7041b70dab6ac212aee, type: 2}
+  - {fileID: 11400000, guid: 5245c784b982dd54a81e3c4acf4e70c5, type: 2}
+  - {fileID: 11400000, guid: c8b93e433aceb3043a793ea46d8bdcaf, type: 2}
+  _inventoryPanel: {fileID: 725999072}
+  _itemInfoPrefab: {fileID: 5900362700406004404, guid: 86df1762ca180e246a48e94cdc92a659, type: 3}
+--- !u!114 &186694794
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 186694792}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: f88e39f468de34142b321113c62a94fc, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _inventoryItemList: []
+--- !u!4 &186694795
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 186694792}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 926.86176, y: 492.9154, z: -13.95443}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_RootOrder: 5
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &431833041
 GameObject:
   m_ObjectHideFlags: 0
@@ -518,7 +584,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!224 &725999072
 RectTransform:
   m_ObjectHideFlags: 0
@@ -760,7 +826,7 @@ RectTransform:
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
-  m_AnchorMax: {x: 1, y: 1}
+  m_AnchorMax: {x: 0, y: 0}
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 20, y: 20}
   m_Pivot: {x: 0.5, y: 0.5}
@@ -946,7 +1012,7 @@ RectTransform:
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
-  m_AnchorMax: {x: 1, y: 1}
+  m_AnchorMax: {x: 0, y: 0}
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 20, y: 20}
   m_Pivot: {x: 0.5, y: 0.5}
@@ -1024,9 +1090,9 @@ RectTransform:
   m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
-  m_AnchorMax: {x: 1, y: 0}
+  m_AnchorMax: {x: 0, y: 0}
   m_AnchoredPosition: {x: 0, y: 0}
-  m_SizeDelta: {x: -17, y: 20}
+  m_SizeDelta: {x: 0, y: 20}
   m_Pivot: {x: 0, y: 0}
 --- !u!114 &1247495492
 MonoBehaviour:
@@ -1125,7 +1191,6 @@ GameObject:
   m_Component:
   - component: {fileID: 1553586390}
   - component: {fileID: 1553586389}
-  - component: {fileID: 1553586391}
   m_Layer: 0
   m_Name: GameManager
   m_TagString: GameController
@@ -1147,14 +1212,6 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   State: 0
   endingScreen: {fileID: 4310544141491983196}
-  _itemList:
-  - {fileID: 11400000, guid: 08c594da8b028ca44b156b2b718a738f, type: 2}
-  - {fileID: 11400000, guid: b793a0ba0d49da840a4eef3ba31165d6, type: 2}
-  - {fileID: 11400000, guid: 9feb9742a522f7041b70dab6ac212aee, type: 2}
-  - {fileID: 11400000, guid: 5245c784b982dd54a81e3c4acf4e70c5, type: 2}
-  - {fileID: 11400000, guid: c8b93e433aceb3043a793ea46d8bdcaf, type: 2}
-  _inventoryPanel: {fileID: 725999072}
-  _itemInfoPrefab: {fileID: 4594275952704546245, guid: 65c9db9d143c23a408eb0dff1c83d30b, type: 3}
 --- !u!4 &1553586390
 Transform:
   m_ObjectHideFlags: 0
@@ -1170,19 +1227,6 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &1553586391
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1553586388}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f88e39f468de34142b321113c62a94fc, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  _inventoryItemList: []
 --- !u!1 &1624610405
 GameObject:
   m_ObjectHideFlags: 0
