+++ Assets/StoryFiles/Story1Chapter1.xml
@@ -8,7 +8,7 @@
 	</DialogueChoices>
 	<DialogueStory>
 		<Node id="1" isRootNode="True"> Start Chapter 1</Node>
-		<Node id="2" item="">Story 1</Node>
+		<Node id="2" item="Bread">Story 1</Node>
 		<Node id="3" isGameOver="True">Story 2 Game Over</Node>
 		<Node id="4" >Story 1 Move Forward</Node>
 		<Node id="5" >Story 1 Continue, Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. </Node>
